Refactor: 코드 구조 변경(Filter.tsx 컴포넌트 분리)
