+++ src/shared/components/filter/Price.tsx
@@ -1,0 +1,10 @@
+const Price = () => {
+  return (
+    <div className="flex w-full flex-col gap-[4px] self-start">
+      <div className="mb-[8px] ml-[32px] self-start text-[20px] font-[500]">가격</div>
+      <div className="ml-[32px]">가격표시공간</div>
+    </div>
+  );
+};
+
+export default Price;
